Move webauthn to a separate crate
In preparation for writing some weabuthn client code
