+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "idelephant"
 version = "0.1.0"
 edition = "2021"
 
+[workspace]
+members = [
+    "webauthn",
+]
 
 [dependencies]
 anyhow = "1.0.95"
@@ -17,17 +21,20 @@ chrono = { version = "0.4.39", features = ["serde"]}
 tracing = "0.1.41"
 tracing-subscriber = {  version = "0.3.19", features = ["env-filter"] }
 rand = "0.9.0"
-base64 = "0.22.1"
+base64 = { workspace = true }
 surrealdb = { version = "2.2.0", features = ["kv-mem", "kv-surrealkv"] }
 tower-sessions = "0.14.0"
 tower-sessions-surrealdb-store = "0.6.0"
-thiserror = "2.0.11"
-ciborium = "0.2.2"
-serde_json = "1.0.138"
 time = "0.3.36"
-sha2 = "0.10.8"
-spki = "0.7.3"
-p256 = "0.13.2"
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+idelephant_webauthn = { path = "webauthn"}
+
+[workspace.dependencies]
+serde_json = "1.0.138"
+base64 = "0.22.1"
+thiserror = "2.0.11"
 
 [dev-dependencies]
 tempfile = "3.15.0"
+
