+++ src/auth.rs
@@ -2,7 +2,6 @@ use crate::error::IdentityError;
 use crate::persistence::Identity;
 use crate::persistence::IdentityState::Active;
 use crate::util::make_token;
-use crate::webauthn::PublicKeyCredentialAuthenticate;
 use crate::AppState;
 use anyhow::{anyhow, Context};
 use axum::extract::State;
@@ -10,6 +9,7 @@ use axum::routing::{get, post};
 use axum::{Json, Router};
 use base64::engine::general_purpose::STANDARD_NO_PAD;
 use base64::Engine;
+use idelephant_webauthn::PublicKeyCredentialAuthenticate;
 use serde::Serialize;
 use serde_json::Value;
 use std::str;
