+++ src/register.rs
@@ -2,7 +2,6 @@ use crate::error::IdentityError;
 use crate::persistence::IdentityState::Allocated;
 use crate::persistence::{Credential, Identity, IdentityState};
 use crate::util::make_token;
-use crate::webauthn::PublicKeyCredentialRegister;
 use crate::AppState;
 use anyhow::anyhow;
 use axum::extract::State;
@@ -11,6 +10,7 @@ use axum::{Json, Router};
 use base64::engine::general_purpose::STANDARD_NO_PAD;
 use base64::Engine;
 use chrono::Utc;
+use idelephant_webauthn::PublicKeyCredentialRegister;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::str;
