+++ webauthn/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "idelephant_webauthn"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.95"
+sha2 = "0.10.8"
+spki = "0.7.3"
+p256 = "0.13.2"
+ciborium = "0.2.2"
+serde_json = { workspace = true }
+base64 = { workspace = true }
+thiserror = { workspace = true }
